[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 1112 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.c
@@ -2,9 +2,9 @@
 
 int main(void)
 {
-    int a, b;
+    long long a, b;
     
-    scanf("%d %d", &a, &b);
-    printf("%d", (a+b) * (a-b));
+    scanf("%lld %lld", &a, &b);
+    printf("%lld", (a+b) * (a-b));
     return (0);
 }
